* include/util_ldap.h (util_ldap_state_t): Fix gcc format string warnings: use a long for connectionTimeout since the code expects a long.
* modules/ldap/util_ldap.c (compare_client_certs): Make static to fix
gcc missing-prototype warning.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@153390 13f79535-47bb-0310-9956-ffa450edef68

+++ include/util_ldap.h
@@ -129,7 +129,7 @@ typedef struct util_ldap_state_t {
     /* cache ald */
     void *util_ldap_cache;
     char *lock_file;           /* filename for shm lock mutex */
-    int connectionTimeout;
+    long  connectionTimeout;
 
 } util_ldap_state_t;
